[GOHW-2]: Add currency functionality

+++ internal/repository/spend/record.go
@@ -4,7 +4,7 @@ import "time"
 
 type SpendRecord struct {
 	ID       int64
-	Price    int64
+	Price    float64
 	Category string
 	DateTime time.Time
 }
